ai_agent/requirements.txt
langgraph>=0.2.0
google-genai>=1.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
streamlit>=1.36.0
